000: split style guide from template
This change adds some features to the repo's style guide:

- folding long content
- info blocks
- multiple authors

It also adds a new markdown template, which uses each element mentioned
in the style guide, but doesn't explain them. This should allow for
easier copy-and-adapt use by authors, as the styles guide's dual,
pre-and-post-rendered view of each element doesn't lend itself to being
copied and then edited inline.

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>