content: supercharging buildkite dynamic pipelines
This change adds a guide that takes the reader through the process of
making a static & known-good Buildkite pipeline dynamic.

It uses the steps contained in a Buildkite blog post to assemble the
known-good starting point and place the reader in a state where the
guide can inform them, without having to deal with too many branching
paths and possibilities.

It uses directories and CUE package names that are compatible with the
simpler static pipeline import guide proposed in
cue-labs/cue-by-example#23. The two guides haven't been explicitly
designed to work together, apart from those naming concerns - but
nothing else stands out as being obviously incompatible.

Closes cue-labs/cue-by-example#21

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>