005_gitlab_ci: use upstream pipeline schema
This amends the GitLab CI guide to use the now-cue-importable upstream
JSON Schema file, which was fixed in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/148555.

The state of the guide proposed here has been tested from scratch and
confirmed as working. A minor change to the process outlined in the
guide was needed (temporarily, I hope & suspect!) to work around the
newly-opened cue-lang/cue#3021, but the guide wasn't amended to reflect
this as it only strikes the module path used when following the
instructions in the context of the guide's specific example repo.

A couple of trivial, unrelated formatting/wording changes are included
in passing.

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>