001_github_actions: fix multiple issues
This updates the GHA guide and fixes its current breakage.

It adds a missing "-p" to "cue import"; works around cue-lang/cue#3462
by noting some versions of the cue command to avoid; and updates
the Schema Store JSONSchema URL to the most recent commit so that the
file downloaded has a $schema field that's acceptable to CUE v0.10.0.

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>