001_github_actions: minimum CUE == v0.11.0-alpha.2
This updates the GitHub Actions guide to make it clear that
v0.11.0-alpha.2 is required for the guide's process to succeed.

This is a combination of cue-lang/cue#3462 being fixed in alpha.2, and
the JSONSchema-to-CUE import having subtle differences which require
annoyingly incompatible workarounds if we were to vary this guide across
CUE versions v0.8.x, v0.9.x, v0.10.x, and v0.11.x. It's far simpler to
insist on alpha.2 and not need any workarounds at all.

A couple of commands are updated to make the guide's testing in
cue-lang/cuelang.org's CI easier. The guide was added to CI by
https://cuelang.org/cl/1201925.

Fixes cue-labs/cue-by-example#34

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>