004_mythic_beasts_dns: relax host constraint
The Mythic Beasts DNS guide is broken by when used with CUE v0.9.x or
later. This updates the guide by removing a MinFields contraint,
allowing it to work with current CUE versions. We can re-add the
constraint when cue-lang/cue#3474 is resolved.

Also: `cue fmt` individual and embedded files; reword a comment.

Fixes cue-labs/cue-by-example#36

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>