005_gitlab_ci: fixes following cuelang.org import
This updates the GitLab CI pipeline guide in line with fixes and
modifications that were required when importing it into cuelang.org's
/docs/draft/ hierarchy in https://cuelang.org/cl/1202059, and later
updated in https://cuelang.org/cl/1203334.

Of particular note is the guide's minimum CUE version of
v0.11.0-alpha.4, which is required both to pick up the fix for
cue-lang/cue#3462 and to opt into the delayed imposition of JSON
Schema's "strict" mode.

Some prose is updated in 006_buildkite_importing_pipelines to reflect a
cue-lang/cue issue that has been fixed since the guide was published.

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>